Swap string values for `app_language` and `select_app_language`.
The string `app_language` is used for the settings item label, and `select_app_language` is used as the title for the language selection dialog. This commit swaps their values in English and Czech to reflect their intended usage and adds `select_app_language` to other languages with a default English value.